Migrated docs to terra-core-docs (#3510)
