+++ packages/terra-core-docs/src/terra-dev-site/doc/form-field/FormField.1.doc.mdx
@@ -1,6 +1,6 @@
 import { Badge } from 'terra-form-field/package.json?dev-site-package';
 
-import FieldExamples from '../example/FieldExamples?dev-site-example';
+import FieldExamples from './example/FieldExamples?dev-site-example';
 
 import FormFieldPropsTable from 'terra-form-field/src/Field?dev-site-props-table';
