+++ packages/terra-core-docs/src/terra-dev-site/test/form-field/FieldHiddenLabel.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Field from '../../../Field';
+import Field from 'terra-form-field';
 import styles from './FormFieldTestCommon.module.scss';
 
 const cx = classNames.bind(styles);
