+++ packages/terra-core-docs/src/terra-dev-site/test/form-field/FieldLabelRequiredErrorHelpText.test.jsx
@@ -1,21 +1,25 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Field from '../../../Field';
+import Field from 'terra-form-field';
 import styles from './FormFieldTestCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
-const FieldLabel = () => (
+const FieldLabelRequiredErrorHelpText = () => (
   <div>
-    <h3>Field - Label</h3>
+    <h3>Field -  Label + Required + In Error + Help Text</h3>
     <Field
-      id="label"
+      id="required-invalid"
       className={cx('form-field')}
       label="Field Label"
+      help="Help Message"
+      error="Error Message"
+      isInvalid
+      required
     >
       <div className={cx('field-content')}>Control Placeholder</div>
     </Field>
   </div>
 );
 
-export default FieldLabel;
+export default FieldLabelRequiredErrorHelpText;
